pyproject.toml
[tool.poetry]
name = "cliper"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
annotated-types = "0.6.0"
anyio = "4.1.0"
argon2-cffi = "23.1.0"
argon2-cffi-bindings = "21.2.0"
attrs = "23.1.0"
automat = "22.10.0"
beautifultable = "1.1.0"
blinker = "1.7.0"
certifi = "2023.11.17"
cffi = "1.16.0"
charset-normalizer = "3.3.2"
click = "8.1.7"
constantly = "23.10.4"
deprecated = "1.2.14"
deprecation = "2.1.0"
environs = "9.5.0"
exceptiongroup = "1.2.0"
filelock = "3.13.1"
flask = "3.0.0"
flask-cors = "4.0.0"
flupy = "1.2.0"
fsspec = "2023.12.1"
ftfy = "6.1.3"
gotrue = "2.1.0"
grpcio = "1.58.0"
h11 = "0.14.0"
httpcore = "0.17.3"
httpx = "0.24.1"
huggingface-hub = "0.19.4"
hyperlink = "21.0.0"
idna = "3.6"
importlib-metadata = "7.0.0"
incremental = "22.10.0"
itsdangerous = "2.1.2"
jinja2 = "3.1.2"
markupsafe = "2.1.3"
marshmallow = "3.20.1"
minio = "7.2.0"
mpmath = "1.3.0"
networkx = "3.2.1"
numpy = "1.26.2"
open-clip-torch = "2.23.0"
packaging = "23.2"
pandas = "2.1.3"
pgvector = "0.1.8"
pillow = "10.1.0"
postgrest = "0.13.0"
protobuf = "4.25.1"
psycopg2-binary = "2.9.9"
pyarrow = "14.0.1"
pycparser = "2.21"
pycryptodome = "3.19.0"
pydantic = "2.5.2"
pydantic-core = "2.14.5"
pymilvus = "2.3.4"
pysondb = "1.6.2"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
pytz = "2023.3.post1"
pyyaml = "6.0.1"
realtime = "1.0.2"
regex = "2023.10.3"
requests = "2.31.0"
safetensors = "0.4.1"
sentencepiece = "0.1.99"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.23"
storage3 = "0.7.0"
strenum = "0.4.15"
supabase = "2.2.1"
supafunc = "0.3.1"
sympy = "1.12"
timm = "0.9.12"
tinydb = "4.8.0"
torch = "2.0.0"
torchvision = "0.15.1"
tqdm = "4.66.1"
twisted = "23.10.0"
typing-extensions = "4.8.0"
tzdata = "2023.3"
ujson = "5.8.0"
urllib3 = "2.1.0"
vecs = "0.4.1"
watchdog = "3.0.0"
wcwidth = "0.2.12"
websockets = "11.0.3"
werkzeug = "3.0.1"
wrapt = "1.16.0"
zipp = "3.17.0"
zope-interface = "6.1"
transformers = "^4.36.1"
clip-interrogator = "^0.6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

